requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
supabase>=2.9.0
python-multipart>=0.0.12
python-jose==3.3.0
python-dotenv>=1.0.0
httpx>=0.23.0
google-generativeai>=0.8.0
google-genai
feedparser
playwright>=1.40.0
pytest-asyncio>=0.21.0
dnspython
requests
